third_party/WebKit/LayoutTests/fast/forms/validity-property.html

<html>
<head>
<title>ValidityState minimal test case</title>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<fieldset name="victim">
<select name="victim"></select>
<input name="victim"/>
<textarea name="victim"></textarea>
<button name="victim"></button>
<output name="victim"></output>
<object name="victim"></object>
<keygen name="victim"></keygen>
</fieldset>

<script>
description('Each form control in this document exposes a validity attribute that returns a live instance of ValidityState.');

function countAttr(o) {
    var a = 0;
    for (pisell in o)
        a++;
    return a;
}

var controls = document.getElementsByName("victim");
for (var i = 0; i < controls.length; ++i) {
    debug('===> Tests for ' + controls[i].tagName);
    shouldBeTrue('"validity" in controls[i]');
    shouldBe('countAttr(controls[i].validity)', '10');
}
</script>

<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
